+++ lib/living_document.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'active_support/core_ext/string/filters' # for `#squish`
-require 'memoist'
+require 'memo_wise'
 require 'stringio'
 
 module LivingDocument ; end